Create hook for fetching companies
